refactor(data-providers-sdk): Rename `metals` to `commodities`
