+++ libs/data_providers_sdk/src/price_data/fetchers/mod.rs
@@ -1,5 +1,5 @@
+pub mod commodities;
 pub mod exchanges;
 pub mod fetch;
 pub mod forex;
-pub mod metals;
 pub mod stock_markets;
